+++ test/python/transpiler/test_direction_mapper.py
@@ -9,7 +9,7 @@
 
 import unittest
 
-from qiskit import QuantumRegister, QuantumCircuit
+from qiskit import ClassicalRegister, QuantumRegister, QuantumCircuit
 from qiskit.transpiler import MapperError
 from qiskit.transpiler.passes import DirectionMapper
 from qiskit.mapper import Coupling
@@ -42,7 +42,7 @@ class TestDirectionMapper(QiskitTestCase):
         self.assertEqual(dag, after)
 
     def test_direction_error(self):
-        """
+        """ The mapping cannot be fixed by direction mapper
          qr0:---------
 
          qr1:---(+)---
@@ -63,7 +63,7 @@ class TestDirectionMapper(QiskitTestCase):
             pass_.run(dag)
 
     def test_direction_correct(self):
-        """
+        """ The CX is in the right direction
          qr0:---(+)---
                  |
          qr1:----.----
@@ -82,16 +82,16 @@ class TestDirectionMapper(QiskitTestCase):
         self.assertEqual(dag, after)
 
     def test_direction_flip(self):
-        """
+        """ Flip a CX
          qr0:----.----
                  |
          qr1:---(+)---
 
          Coupling map: [0] -> [1]
 
-         qr0:---(+)---
-                 |
-         qr1:----.----
+         qr0:-[H]-(+)-[H]--
+                   |
+         qr1:-[H]--.--[H]--
         """
         qr = QuantumRegister(2, 'qr')
         circuit = QuantumCircuit(qr)
@@ -105,6 +105,44 @@ class TestDirectionMapper(QiskitTestCase):
         expected.cx(qr[0], qr[1])
         expected.h(qr[0])
         expected.h(qr[1])
+
+        pass_ = DirectionMapper(coupling)
+        after = pass_.run(dag)
+
+        self.assertEqual(circuit_to_dag(expected), after)
+
+    def test_flip_with_measure(self):
+        """
+         qr0: -(+)-[m]-
+                |   |
+         qr1: --.---|--
+                    |
+         cr0: ------.--
+
+         Coupling map: [0] -> [1]
+
+         qr0: -[H]--.--[H]-[m]-
+                    |       |
+         qr1: -[H]-(+)-[H]--|--
+                            |
+         cr0: --------------.--
+        """
+        qr = QuantumRegister(2, 'qr')
+        cr = ClassicalRegister(1, 'cr')
+
+        circuit = QuantumCircuit(qr, cr)
+        circuit.cx(qr[1], qr[0])
+        circuit.measure(qr[0], cr[0])
+        coupling = Coupling({0: [1]})
+        dag = circuit_to_dag(circuit)
+
+        expected = QuantumCircuit(qr, cr)
+        expected.h(qr[0])
+        expected.h(qr[1])
+        expected.cx(qr[0], qr[1])
+        expected.h(qr[0])
+        expected.h(qr[1])
+        expected.measure(qr[0], cr[0])
 
         pass_ = DirectionMapper(coupling)
         after = pass_.run(dag)